+++ install-stubs/Orchid/Filters/RoleFilter.php
@@ -7,7 +7,7 @@ namespace App\Orchid\Filters;
 use Orchid\Screen\Field;
 use Orchid\Platform\Models\Role;
 use Orchid\Screen\Fields\Select;
-use Orchid\Platform\Filters\Filter;
+use Orchid\Filters\Filter;
 use Illuminate\Database\Eloquent\Builder;
 
 class RoleFilter extends Filter