+++ src/Access/UserInterface.php
@@ -6,4 +6,11 @@ namespace Orchid\Access;
 
 interface UserInterface
 {
+    /**
+     * Returns all role for the user.
+     *
+     * @return \IteratorAggregate
+     */
+    public function getRoles();
+
 }