+++ src/Platform/Models/Role.php
@@ -7,12 +7,12 @@ namespace Orchid\Platform\Models;
 use Orchid\Access\RoleAccess;
 use Orchid\Access\RoleInterface;
 use Illuminate\Database\Eloquent\Model;
-use Orchid\Platform\Traits\FilterTrait;
+use Orchid\Filters\Filterable;
 use Orchid\Platform\Traits\MultiLanguageTrait;
 
 class Role extends Model implements RoleInterface
 {
-    use RoleAccess, FilterTrait, MultiLanguageTrait;
+    use RoleAccess, Filterable, MultiLanguageTrait;
 
     /**
      * @var string