+++ src/Press/Http/Filters/StatusFilter.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Orchid\Press\Http\Filters;
 
 use Orchid\Screen\Field;
-use Orchid\Platform\Filters\Filter;
+use Orchid\Filters\Filter;
 use Orchid\Screen\Fields\RadioButtons;
 use Illuminate\Database\Eloquent\Builder;