+++ src/Screen/Layouts/Selection.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Orchid\Screen\Layouts;
 
 use Orchid\Screen\Repository;
-use Orchid\Platform\Filters\Filter;
+use Orchid\Filters\Filter;
 
 /**
  * Class Selection.