+++ src/Support/helpers.php
@@ -71,7 +71,7 @@ if (! function_exists('is_sort')) {
      */
     function is_sort($property = null)
     {
-        $filter = new \Orchid\Platform\Filters\HttpFilter();
+        $filter = new \Orchid\Filters\HttpFilter();
 
         return $filter->isSort($property);
     }
@@ -86,7 +86,7 @@ if (! function_exists('get_sort')) {
      */
     function get_sort($property)
     {
-        $filter = new \Orchid\Platform\Filters\HttpFilter();
+        $filter = new \Orchid\Filters\HttpFilter();
 
         return $filter->getSort($property);
     }
@@ -101,7 +101,7 @@ if (! function_exists('get_filter')) {
      */
     function get_filter($property)
     {
-        $filter = new \Orchid\Platform\Filters\HttpFilter();
+        $filter = new \Orchid\Filters\HttpFilter();
 
         return $filter->getFilter($property);
     }
@@ -135,7 +135,7 @@ if (! function_exists('revert_sort')) {
      */
     function revert_sort($property)
     {
-        $filter = new \Orchid\Platform\Filters\HttpFilter();
+        $filter = new \Orchid\Filters\HttpFilter();
 
         return $filter->revertSort($property);
     }