+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Chart min heating auto
+- Cache browser views
+
+### Fixed
+- Repeated run test to create resource links
+
+### Deprecated
+- `Orchid\Platform\Filters\*` use `Orchid\Filters\Filter\*`
+- `AttachTrait` use `Orchid\Attachment\Attachable`
+- `FilterTrait` use `Orchid\Filters\Filterable`
+
 ## 4.6.1 - 2019-04-16
 
 ### Changed